pyproject.toml
[tool.poetry]
name = "spa-add-comments"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
pillow = "^10.3.0"
captcha = "^0.5.0"
django-simple-captcha = "^0.6.0"
lxml = "^5.2.2"
lxml-html-clean = "^0.1.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
